chore: split columnview columns
